Fixed SonarCloud issues, removed duplicate code and unused macro

+++ src/raspberrypi/controllers/scsi_controller.cpp
@@ -1065,7 +1065,7 @@ void ScsiController::ProcessMessage()
 
 			// IDENTIFY
 			if (data >= 0x80) {
-				identified_lun = data & 0x1F;
+				identified_lun = (int)data & 0x1F;
 				LOGTRACE("Message code IDENTIFY $%02X, LUN %d selected", data, identified_lun)
 			}
